refactor(constructors): rename operator -> constructor across I-0132 (CLOACI-I-0132)
Nomenclature: the reusable, parameterized, WASM-packaged factory that becomes a
cloacina primitive is a 'constructor' (it constructs a primitive; aligns with
cloacina's Runtime constructor vocabulary + fidius's configure constructor) — not
'operator' (Airflow's word, and an Airflow operator IS a task, while ours is a
factory for any of task/trigger/accumulator/reactor).

Renamed in lockstep (guest + host) so the fidius interface-hash gate still matches:
- crate cloacina-operator-contract -> cloacina-constructor-contract
- #[operator] -> #[constructor]; operator_attr.rs -> constructor_attr.rs
- OperatorManifest/Error -> Constructor*; Task/TriggerOperator -> *Constructor;
  WasmTask/TriggerOperator -> WasmTask/TriggerConstructor
- operator_loader.rs -> constructor_loader.rs; load_*_operator -> load_*_constructor
- operator.json -> constructor.json; feature operators-wasm -> constructors-wasm
- examples/operator-contract -> examples/constructor-contract (+ fixtures); tests
- registry closure param 'constructor' -> 'factory' (register_task/register_trigger)
- I-0132 initiative + open tasks retitled (id slugs preserved)

Verified: cargo check --workspace clean; no wasmtime by default; fmt clean; all
WASM tests pass post-rename (loader 3, trigger 6, macro 4, contract 4) — the macro
test proves the interface-hash gate still matches. Historical wasm-operator-spike
left intact.

CLOACI-I-0132.
