+++ crates/cloacina-macros/src/lib.rs
@@ -45,7 +45,7 @@
 //! ```
 
 pub(crate) mod computation_graph;
-mod operator_attr;
+mod constructor_attr;
 pub(crate) mod packaged_workflow;
 mod reactor_attr;
 mod registry;
@@ -159,8 +159,8 @@ pub fn reactor(args: TokenStream, input: TokenStream) -> TokenStream {
     reactor_attr::reactor_attr(args, input)
 }
 
-/// Author a WASM **operator** in clean form; generate the raw fidius contract
-/// (sync trait impl + config + JSON wire + the `operator.json` manifest data)
+/// Author a WASM **constructor** in clean form; generate the raw fidius contract
+/// (sync trait impl + config + JSON wire + the `constructor.json` manifest data)
 /// the loader consumes (CLOACI-T-0826).
 ///
 /// Applied to a struct whose fields are `#[config]` (bound once per instance at
@@ -168,16 +168,16 @@ pub fn reactor(args: TokenStream, input: TokenStream) -> TokenStream {
 /// from the task context). The author writes ONLY the body method (`execute` for
 /// `kind = task`); the macro emits the `#[plugin_interface]` sync trait, the
 /// `#[plugin_impl(config = …)]` impl + `configure` hook, and a
-/// `pub fn __operator_manifest() -> OperatorManifest`.
+/// `pub fn __constructor_manifest() -> ConstructorManifest`.
 ///
 /// ```rust,ignore
-/// #[operator(kind = task, name = "prefix", version = "0.1.0")]
+/// #[constructor(kind = task, name = "prefix", version = "0.1.0")]
 /// struct Prefix {
 ///     #[config] prefix: String,
 ///     #[param(required)] name: String,
 /// }
 /// impl Prefix {
-///     fn execute(&self) -> Result<(), OperatorError> {
+///     fn execute(&self) -> Result<(), ConstructorError> {
 ///         self.set("result", format!("{}{}", self.prefix, self.name));
 ///         Ok(())
 ///     }
@@ -187,8 +187,8 @@ pub fn reactor(args: TokenStream, input: TokenStream) -> TokenStream {
 /// Only `kind = task` is code-generated today; the sibling kinds (trigger /
 /// accumulator / reactor) map onto the same shape and are a noted continuation.
 #[proc_macro_attribute]
-pub fn operator(args: TokenStream, input: TokenStream) -> TokenStream {
-    operator_attr::operator_attr(args, input)
+pub fn constructor(args: TokenStream, input: TokenStream) -> TokenStream {
+    constructor_attr::constructor_attr(args, input)
 }
 
 /// Define a passthrough accumulator (socket-only, no event loop).
