+++ examples/constructor-contract/host/Cargo.toml
@@ -1,0 +1,27 @@
+# CLOACI-T-0822 constructor contract — host side (tests).
+#
+# Loads the task-constructor-fixture component via fidius `load_wasm_configured`,
+# proves (1) the constructor MANIFEST round-trips (produced -> package -> read) and
+# carries the primitive kind the loader switches on, and (2) the configured
+# TASK constructor loads + invokes across the sandbox with a serialized context.
+#
+# Standalone crate, EXCLUDED from the cloacina workspace (under examples/* plus
+# the empty [workspace] table). fidius-host `wasm` feature pulls wasmtime.
+[workspace]
+
+[package]
+name = "constructor-contract-host"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+fidius-host = { version = "0.5.4", features = ["wasm"] }
+fidius-macro = "0.5.4"
+fidius-core = "0.5.4"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+constructor-contract = { path = "../constructor-contract" }
+
+[dev-dependencies]
+tempfile = "3"
