+++ examples/constructor-contract/host/src/lib.rs
@@ -1,0 +1,2 @@
+// CLOACI-T-0822 constructor contract — host side (empty lib; the proof is the
+// integration test under tests/).
