+++ examples/constructor-contract/task-constructor-fixture/Cargo.toml
@@ -1,0 +1,35 @@
+# CLOACI-T-0822 constructor contract — TASK constructor guest fixture.
+#
+# A SYNCHRONOUS task constructor authored with the fidius macros and built to a
+# wasm32-wasip2 component (the T-0821 recipe). Implements the `TaskConstructor`
+# contract: JSON(TaskInvocation) in -> JSON(TaskOutcome) out. Config (the
+# instance's `prefix`) binds ONCE via the macro-emitted `fidius-configure`
+# export; `execute` then reads it without the caller re-passing it.
+#
+# Standalone crate, EXCLUDED from the cloacina workspace (under examples/*,
+# plus the empty [workspace] table below) so `cargo build --target
+# wasm32-wasip2` stays self-contained.
+[workspace]
+
+[package]
+name = "task-constructor-fixture"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+fidius-guest = "0.5.4"
+fidius-macro = "0.5.4"
+wit-bindgen = "0.44"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+# Shared, language-neutral contract types (wire structs). serde-only → wasm-safe.
+constructor-contract = { path = "../constructor-contract" }
+
+[profile.release]
+opt-level = "s"
+lto = true
+strip = true
