+++ examples/constructor-contract/task-constructor-macro-fixture/Cargo.toml
@@ -1,0 +1,54 @@
+# CLOACI-T-0826 — `#[constructor]`-authored TASK constructor guest fixture.
+#
+# The macro counterpart to `task-constructor-fixture`: the author writes the CLEAN
+# constructor form (a struct with #[config]/#[param] fields + a single `execute`
+# body) and `#[constructor(kind = task, ...)]` generates the raw fidius contract
+# (the sync `TaskConstructor` trait + `#[plugin_impl(config = ...)]` + `configure` +
+# the JSON wire) plus a `__constructor_manifest()` fn. Built to a wasm32-wasip2
+# component, it loads + runs through the cloacina `constructors-wasm` loader exactly
+# like the hand-written fixture.
+#
+# The fidius guest glue is emitted under `#[cfg(target_arch = "wasm32")]`, so the
+# crate ALSO compiles on the host with only the struct + `__constructor_manifest()`
+# — that is how the `emit_manifest` bin materializes `constructor.json` for the
+# end-to-end test without the wasm-only exports.
+#
+# Standalone crate, EXCLUDED from the cloacina workspace (under examples/* plus
+# the empty [workspace] table) so the wasm build stays self-contained.
+[workspace]
+
+[package]
+name = "task-constructor-macro-fixture"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[lib]
+# cdylib → the wasm component; rlib → so `emit_manifest` (host) can link it.
+crate-type = ["cdylib", "rlib"]
+
+[[bin]]
+name = "emit_manifest"
+path = "src/bin/emit_manifest.rs"
+
+[dependencies]
+# The `#[constructor]` authoring macro under test (path dep; proc-macro runs on host).
+cloacina-macros = { path = "../../../crates/cloacina-macros" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+# Shared, language-neutral contract types (wire structs + manifest + ConstructorError).
+# serde-only → wasm-safe AND host-safe.
+constructor-contract = { path = "../constructor-contract" }
+
+# fidius guest SDK is needed ONLY for the wasm32 component build (the macro's
+# guest glue is `#[cfg(target_arch = "wasm32")]`). Target-gating it keeps the
+# host build (the `emit_manifest` bin) free of wasm-only deps.
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+fidius-guest = "0.5.4"
+fidius-macro = "0.5.4"
+wit-bindgen = "0.44"
+
+[profile.release]
+opt-level = "s"
+lto = true
+strip = true
