+++ examples/constructor-contract/trigger-constructor-fixture/Cargo.toml
@@ -1,0 +1,36 @@
+# CLOACI-T-0824 constructor contract — TRIGGER constructor guest fixture.
+#
+# A SYNCHRONOUS trigger constructor authored with the fidius macros and built to a
+# wasm32-wasip2 component (the T-0821 recipe). Implements the `TriggerConstructor`
+# contract: JSON(TriggerInvocation) in -> JSON(PollOutcome) out. Config (whether
+# this instance should fire, and the message it fires with) binds ONCE via the
+# macro-emitted `fidius-configure` export; `poll` then reads it without the
+# caller re-passing it — proving config-bound Fire/Skip across the sandbox.
+#
+# Standalone crate, EXCLUDED from the cloacina workspace (under examples/*, plus
+# the empty [workspace] table below) so `cargo build --target wasm32-wasip2`
+# stays self-contained.
+[workspace]
+
+[package]
+name = "trigger-constructor-fixture"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+fidius-guest = "0.5.4"
+fidius-macro = "0.5.4"
+wit-bindgen = "0.44"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+# Shared, language-neutral contract types (wire structs). serde-only → wasm-safe.
+constructor-contract = { path = "../constructor-contract" }
+
+[profile.release]
+opt-level = "s"
+lto = true
+strip = true
