--- examples/operator-contract/host/Cargo.toml
@@ -1,27 +1,0 @@
-# CLOACI-T-0822 operator contract — host side (tests).
-#
-# Loads the task-operator-fixture component via fidius `load_wasm_configured`,
-# proves (1) the operator MANIFEST round-trips (produced -> package -> read) and
-# carries the primitive kind the loader switches on, and (2) the configured
-# TASK operator loads + invokes across the sandbox with a serialized context.
-#
-# Standalone crate, EXCLUDED from the cloacina workspace (under examples/* plus
-# the empty [workspace] table). fidius-host `wasm` feature pulls wasmtime.
-[workspace]
-
-[package]
-name = "operator-contract-host"
-version = "0.0.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-fidius-host = { version = "0.5.4", features = ["wasm"] }
-fidius-macro = "0.5.4"
-fidius-core = "0.5.4"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-operator-contract = { path = "../operator-contract" }
-
-[dev-dependencies]
-tempfile = "3"
