--- examples/operator-contract/host/src/lib.rs
@@ -1,2 +1,0 @@
-// CLOACI-T-0822 operator contract — host side (empty lib; the proof is the
-// integration test under tests/).
