--- examples/operator-contract/task-operator-macro-fixture/Cargo.toml
@@ -1,54 +1,0 @@
-# CLOACI-T-0826 — `#[operator]`-authored TASK operator guest fixture.
-#
-# The macro counterpart to `task-operator-fixture`: the author writes the CLEAN
-# operator form (a struct with #[config]/#[param] fields + a single `execute`
-# body) and `#[operator(kind = task, ...)]` generates the raw fidius contract
-# (the sync `TaskOperator` trait + `#[plugin_impl(config = ...)]` + `configure` +
-# the JSON wire) plus a `__operator_manifest()` fn. Built to a wasm32-wasip2
-# component, it loads + runs through the cloacina `operators-wasm` loader exactly
-# like the hand-written fixture.
-#
-# The fidius guest glue is emitted under `#[cfg(target_arch = "wasm32")]`, so the
-# crate ALSO compiles on the host with only the struct + `__operator_manifest()`
-# — that is how the `emit_manifest` bin materializes `operator.json` for the
-# end-to-end test without the wasm-only exports.
-#
-# Standalone crate, EXCLUDED from the cloacina workspace (under examples/* plus
-# the empty [workspace] table) so the wasm build stays self-contained.
-[workspace]
-
-[package]
-name = "task-operator-macro-fixture"
-version = "0.0.0"
-edition = "2021"
-publish = false
-
-[lib]
-# cdylib → the wasm component; rlib → so `emit_manifest` (host) can link it.
-crate-type = ["cdylib", "rlib"]
-
-[[bin]]
-name = "emit_manifest"
-path = "src/bin/emit_manifest.rs"
-
-[dependencies]
-# The `#[operator]` authoring macro under test (path dep; proc-macro runs on host).
-cloacina-macros = { path = "../../../crates/cloacina-macros" }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-# Shared, language-neutral contract types (wire structs + manifest + OperatorError).
-# serde-only → wasm-safe AND host-safe.
-operator-contract = { path = "../operator-contract" }
-
-# fidius guest SDK is needed ONLY for the wasm32 component build (the macro's
-# guest glue is `#[cfg(target_arch = "wasm32")]`). Target-gating it keeps the
-# host build (the `emit_manifest` bin) free of wasm-only deps.
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-fidius-guest = "0.5.4"
-fidius-macro = "0.5.4"
-wit-bindgen = "0.44"
-
-[profile.release]
-opt-level = "s"
-lto = true
-strip = true
